Implement bind elaboration in the hierarchy pass
This is surprisingly finicky, because of the difference between bind
directives that target modules and those that target specific
instances of those modules (i.e. specific cells).

In the first case, we can just operate on the module in place. In the
second, we have to make fresh copies of the target module and the
"path of cells" from the bind directive down to it. For instance, in
something like this:

    bind $root.foo.bar_i.baz_i qux bound_i (.*);

(where bar_i and baz_i are instances of modules called bar and baz,
respectively), we need to make a fresh clone of baz (let's call it
baz1) into which we insert an instance of qux. But we can't just
modify bar to re-type it's baz_i cell from baz to baz1: what if there
are other instances of bar in the design? So we then need to make a
fresh clone of bar (let's call it bar1). Since foo here refers to the
module, rather than a particular instance, we *can* now modify foo to
re-type it's bar_i cell from bar to bar1 (phew!).

To avoid some extra un-needed renames, we actually count the number of
times each cell type appears in the design (see get_cell_type_counts).
If the design in the example above didn't actually use bar except as
bar_i in module foo, we actually operate in place rather than
renaming.